casi terminado

+++ App/Services/ExperienciaService.php
@@ -8,7 +8,10 @@ class ExperienciaService
 {
     public function deleteExperiencia($id)
     {
-        $experiencia = Experiencia::find($id);
+        $experiencia = Experiencia::findOrFail($id);
+        if (!$experiencia) {
+           throw new \Exception('Experiencia no encontrada');
+        }
         $experiencia->delete();
 
     }
